Change types of position variables from float to int as you can't have part of a pixel

+++ Headers/Bullet.h
@@ -7,7 +7,7 @@ public:
 	void move(const float dt, bool move_down) override;
 	void death() override;
 	void onCollision() override;
-	Bullet(float xpos_ = 0, float ypos_ = 0, bool collides_ = false, Game* game_ = nullptr);
+	Bullet(int xpos_ = 0, int ypos_ = 0, bool collides_ = false, Game* game_ = nullptr);
 
 	bool detectCollision(Entity& other) override;
 };
