+++ Headers/TextManager.h
@@ -12,11 +12,11 @@ private:
 	FontManager* fontManager;
 
 public:
-	void createText(const std::string& name, const std::string& fontName, int size, sf::Color color, std::string string, float xpos, float ypos);
+	void createText(const std::string& name, const std::string& fontName, int size, sf::Color color, std::string string, int xpos, int ypos);
 	void updateTextString(const std::string& textName, std::string string);
 	void updateTextColor(const std::string& textName, sf::Color color);
 	void updateTextSize(const std::string& textName, int size);
-	void updateTextPosition(const std::string& textName, float xpos, float ypos);
+	void updateTextPosition(const std::string& textName, int xpos, int ypos);
 	void updateTextFont(const std::string& textName, const std::string& fontName);
 
 	// Translate an id into a reference
